#218 move messages to a seperate package
